:recycle: refactor: Definindo sidebar como componete global
